+++ packages/keyring/src/pair/index.ts
@@ -6,6 +6,7 @@ import { Keypair, KeypairType } from '@polkadot/util-crypto/types';
 import { KeyringPair, KeyringPair$Json, KeyringPair$Meta } from '../types';
 import { PairInfo } from './types';
 
+import { u8aToHex } from '@polkadot/util';
 import { naclKeypairFromSeed as naclFromSeed, naclSign, naclVerify, schnorrkelKeypairFromSeed as schnorrkelFromSeed, schnorrkelSign, schnorrkelVerify } from '@polkadot/util-crypto';
 
 import { encodeAddress } from '../address';