+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# 0.38.1
+
+- Align with current substrate master for sr25519 keys and hard/soft derivation
+- Swap default keyring to sr25519
+
 # 0.37.1
 
 - Swap to publishing -beta.x on merge (non-breaking testing)